+++ internal/integration_test/vs/runtime.go
@@ -82,7 +82,7 @@ func (m *wazeroModule) Memory() []byte {
 func (r *wazeroRuntime) log(_ context.Context, mod api.Module, stack []uint64) {
 	offset, byteCount := uint32(stack[0]), uint32(stack[1])
 
-	buf, ok := mod.Memory().Read(offset, byteCount)
+	buf, ok := mod.Memory().Read(offset, uint64(byteCount))
 	if !ok {
 		panic("out of memory reading log buffer")
 	}
